// api/model/mediaType.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Photobooth API
 * The Photobooth Backend API
 * OpenAPI spec version: 1.0
 */

/**
 * The type of the media file
 */
export type MediaType = typeof MediaType[keyof typeof MediaType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MediaType = {
  ORIGINAL: 'ORIGINAL',
  PROCESSED: 'PROCESSED',
  VIDEO: 'VIDEO',
} as const;