// api/model/sessionStatus.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Photobooth API
 * The Photobooth Backend API
 * OpenAPI spec version: 1.0
 */

/**
 * The current status of the session
 */
export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SessionStatus = {
  INIT: 'INIT',
  SELECTING: 'SELECTING',
  COUNTDOWN: 'COUNTDOWN',
  CAPTURING: 'CAPTURING',
  COMPLETED: 'COMPLETED',
} as const;